Updated ui creation of google search
* Moved createGoogleUIFromHTMLDoc functions from content.js and popup.js to utils.js to simplify UI creation and reduce clutter.
* Background.js has been changed accordingly. Now the imported document is parsed in background.js instead of popup.js and content.js. Then the necessary data is converted to JSON object and sent. In this way, we avoid getting cloning errors when sending these items. Also, the information stored for each word becomes much smaller for google searches.
